Fix #1: regression in a92e585 causing 'follow' to malfunction
When dropping the QNX #ifdefs in 2016 a subtle but fatal regression was
introduced, causing keywords as 'follow' to malfunction:

    > exit
    You are standing at the end of a road before a small brick
    building.  Around you is a forest.  A small stream flows out
    of the building and down a gully.
    > follow stream
    I don't know how to apply that word here

The refactor in a92e585 managed to map rindex() to strchr() instead of
strrchr() in database.c.  This likely happened because of cut-and-paste
from other source files with *almost* the same #define mappings.  Very
subtle and very annoying.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>